% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_chromosomes.R
\name{all_chromosomes}
\alias{all_chromosomes}
\title{Compile all chromosome lists into data frames.}
\usage{
all_chromosomes(compiled_SNP_data)
}
\arguments{
\item{compiled_SNP_data}{some A list containing the annotated data of each chromosome.}
}
\value{
A list of five components.
\describe{
 \item{introns}{A data frame containing all the stop and start codons of all the SNPs located in introns.}
 \item{exons}{A data frame containing all the stop and start codons, gene location, and genomic region location of all the SNPs in exons}
 \item{stream}{A data frame containing all the start and stop codons for the SNPs in the intergenic regions.}
 \item{genome_with_SNP}{The orginal "allgenes" genome data frame with an additional column containing the SNP count for each region.}
 \item{stream_with_SNP}{A data frame containing the start and stop codon for each intergenic region and the SNP count for each region.}
}
}
\description{
This function takes the annotated SNP data for each chromosome and compiles all the data into five data frames containing
the entire genome, a table with the intergenic data, and tables containing SNPs in exons, introns, and intergenic regions.
}

