% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_ecotype_SNP_count.R
\name{all_ecotype_SNP_count}
\alias{all_ecotype_SNP_count}
\title{Compile all chromosome lists into data frames.}
\usage{
all_ecotype_SNP_count(genome_file, gene_ids, gene_directory, ecotype_list,
  chr_num)
}
\arguments{
\item{compiled_SNP_data}{some A list containing the annotated data of each chromosome.}
}
\value{
A data frame.
}
\description{
Description
}

