% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/background_impact_rates.R
\name{background_impact_rate}
\alias{background_impact_rate}
\title{Title}
\usage{
background_impact_rate(all_data, impact_list, chr_num)
}
\arguments{
\item{GeneIDs}{some stuff}

\item{genome}{}
}
\value{
list
\describe{
 \item{SNP_introns}{stuff }
 \item{SNP_exons}{}
 \item{SNP_stream}{}
 \item{stream_withSNP}{}
 \item{chr_withSNP}{}
}
}
\description{
Description
}

