% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_snp_rate.R
\name{calculate_snp_rate}
\alias{calculate_snp_rate}
\title{Title}
\usage{
calculate_snp_rate(GeneIDs, prepared_genome)
}
\arguments{
\item{GeneIDs}{some stuff}

\item{genome}{}
}
\value{
list
\describe{
 \item{SNP_introns}{stuff }
 \item{SNP_exons}{}
 \item{SNP_stream}{}
 \item{stream_withSNP}{}
 \item{chr_withSNP}{}
}
}
\description{
Description
}

