% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_background_rate.R
\name{compare_background_rate}
\alias{compare_background_rate}
\title{Compare SNP Rates in Gene Lists to Background Rates}
\usage{
compare_background_rate(gene_list, all_IDs, genome, rep_num)
}
\arguments{
\item{gene_list}{A data frame containing the a list of genes to be tested in each column.}

\item{all_IDs}{A list of all genes in the genome.}

\item{genome}{The genome of the species with the snp count of each gene region.}

\item{rep_num}{The number of simulations replications used to estimate the background rateß.}
}
\value{
A list with two elements:
\describe{
 \item{all_trails}{A data frame containing all of the simulated background rates for each gene list.}
 \item{rates}{A table containing the true rates and simulated rates for each gene list.  There are
 three simulated rates: rates including all genes, rates including only TEs, and rates containing
 only genes without TEs.}
}
}
\description{
This function both calculates that SNP rate per nucleotide in specific gene lists and
simulates background rates for comparision to each gene list.  The background rates are simulated
using random lists of the same size as the gene lists and are simulated using three sets of
genes: all genes, only TEs, and all genes but TEs.
}

