% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do_one_chromosome.R
\name{do_one_chromosome}
\alias{do_one_chromosome}
\title{Annotate Genome with SNPs for a Single Chromosome}
\usage{
do_one_chromosome(chr_gene, chr_SNP, w = NULL, do_stream = TRUE)
}
\arguments{
\item{chr_gene}{A data frame containing the genome of the species for one chromosome.
Each row of data frame represents a specific section of the genome.  The data
should at least contain the listed columns:
\describe{
 \item{Region}{Labeled either "exon", "CDS", "5UTR", or "3UTR".}
 \item{Start}{The begining of the genomic region.}
 \item{Stop}{The end of the genomic region.}
 \item{GeneID}{The name of the gene to which the region belongs}
 \item{direction}{A "+" or "-" indicating the strand on which the gene segment is located}
}}

\item{chr_SNP}{A data frame containing the start and stop of each SNP in the specific
chromosome of the genome. The start codon should be labled "Start," and the stop codon should be labeled "Stop."}
}
\value{
A list with five components:
\describe{
 \item{SNP_introns}{A data frame containing all the SNPs in introns}
 \item{SNP_exons}{A data frame containing all the SNPs in exons. The data frame also contains information on the gene name and genomic region in
 which the SNP is found.}
 \item{SNP_stream}{A data frame containing all the SNPs in the intergenic regions}
 \item{stream_withSNP}{Contains the start and stop codon of each intergenic region and the number of SNPs found int the region.}
 \item{chr_withSNP}{The orginal "chr_gene" data frame with an additional column containing the number of SNPs in each gene region.}
}
}
\description{
This function annotates the genome of one chromosome with SNP count and produces data frames containing SNPs in exons, introns
and intergenic regions of the chromosome.
}

