% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impact_list_rates.R
\name{impact_list_rates}
\alias{impact_list_rates}
\title{Title}
\usage{
impact_list_rates(all_data, impact_list, gene_list, chr_num)
}
\arguments{
\item{GeneIDs}{some stuff}

\item{genome}{}
}
\value{
list
\describe{
 \item{SNP_introns}{stuff }
 \item{SNP_exons}{}
 \item{SNP_stream}{}
 \item{stream_withSNP}{}
 \item{chr_withSNP}{}
}
}
\description{
Description
}

