% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_genome.R
\name{prepare_genome}
\alias{prepare_genome}
\title{Prepare Genome for the Functions}
\usage{
prepare_genome(genome)
}
\arguments{
\item{genome}{}
}
\value{
The edited genome with with the GeneID column containing only
the names of the gene with no extra characters.
}
\description{
Description
}

